Project initializing
The preparation phase is done. It checks the project structure and config validity. However I don't like how the verification of the warehouse and the config is done. I suggest later to use classes as the Templates. For now need to be careful when adding new lines to config, as one must add them both in `verify_config_file()` and `create_config()`. Also `repair_config_file` has no real implementation.